deepen: inline analysis-depth scoring into cached-summary-delivery
getAnalysisDepthScore was extracted into its own module solely for
testability — its only production caller was calculateCompositeScore.
The separation created a published interface (ANALYSIS_DEPTH_BONUSES,
AnalysisDepthInput) that existed only to satisfy tests, not callers.

Move the scoring function inline as private implementation. Delete
analysis-depth.ts and its 221-line test file. The 28 calculateCompositeScore
tests already cover the depth signal at the real interface.

ADR: docs/adr/0002-inline-analysis-depth-scoring.md
